fixed quanitity error

+++ src/java/controller/UpdateCartServlet.java
@@ -46,7 +46,7 @@ public class UpdateCartServlet extends HttpServlet {
                 String x = items.nextElement();
                 if (x.charAt(0) == '_') {
                     int pid = Integer.parseInt(x.substring(1));
-                    int qty = Integer.parseInt(request.getParameter(x));
+                    double qty = Double.parseDouble(request.getParameter(x));
                     if (cart.getItem(pid) != null) {
                         cart.updateItem(pid, qty);
                     }
